Update handling of working directory for the IMPACT lattice runner.

+++ phyutil/phylib/model/impact.py
@@ -36,22 +36,25 @@ _TEMP_DIRECTORY_SUFFIX = "_impact"
 _LOGGER = logging.getLogger(__name__)
 
 
-def run(lattice, **kwargs):
+def run_lattice(lattice, **kwargs):
     """Convenience method to build result with specified configuration.
 
     :param lattice: IMPACT Lattice object
     :param config: machine configuration
-    :param settings: dictionary of machine settings
     :param data_dir: path of directory containing IMPACT data files
     :param work_dir: path of directory for execution of IMPACT
     :return: Result instance
     """
-    result_factory = ResultFactory(lattice, **kwargs)
+    impact = kwargs.get("impact", "FRIB")
 
-    return result_factory.build()
+    lattice_runner = LatticeRunner(lattice, **kwargs)
+    directory = lattice_runner.run()
+
+    return build_result(impact=impact, directory=directory, keep=True)
 
 
-class ResultFactory(object):
+
+class LatticeRunner(object):
     """A factory to run IMPACT and get the resulting model."""
 
     def __init__(self, lattice, **kwargs):
@@ -59,7 +62,6 @@ class ResultFactory(object):
 
         :param lattice: IMPACT Lattice object
         :param config: machine configuration
-        :param settings: dictionary of machine settings
         :param data_dir: path of directory containing IMPACT data files
         :param work_dir: path of directory for execution of IMPACT
         """
@@ -68,7 +70,7 @@ class ResultFactory(object):
         if "config" in kwargs:
             self.config = kwargs.get("config")
         else:
-            self.cofig = cfg.config
+            self.config = cfg.config
         
         self.data_dir = kwargs.get("data_dir", None)
         self.work_dir = kwargs.get("work_dir", None)
@@ -81,7 +83,7 @@ class ResultFactory(object):
     @lattice.setter
     def lattice(self, lattice):
         if not isinstance(lattice, Lattice):
-            raise TypeError("ResultFactory: 'lattice' property much be type Lattice")
+            raise TypeError("LatticeRunner: 'lattice' property much be type Lattice")
         self._lattice = lattice
 
 
@@ -92,7 +94,7 @@ class ResultFactory(object):
     @data_dir.setter
     def data_dir(self, data_dir):
         if (data_dir != None) and not isinstance(data_dir, basestring):
-            raise TypeError("ResultFactory: 'data_dir' property much be type string or None")
+            raise TypeError("LatticeRunner: 'data_dir' property much be type string or None")
         self._data_dir = data_dir
 
 
@@ -103,48 +105,58 @@ class ResultFactory(object):
     @work_dir.setter
     def work_dir(self, work_dir):
         if (work_dir != None) and not isinstance(work_dir, basestring):
-            raise TypeError("ResultFactory: 'work_dir' property much be type string or None")
+            raise TypeError("LatticeRunner: 'work_dir' property much be type string or None")
         self._work_dir = work_dir
 
 
+    @property
+    def config(self):
+        return self._config
+
+    @config.setter
+    def config(self, config):
+        if not isinstance(config, (cfg.Configuration)):
+            raise TypeError("LatticeRunner: 'config' property must be type Configuration")
+        self._config = config
+
+
     def _get_config_impact_exe(self):
-        if cfg.config.has_default(CONFIG_IMPACT_EXE_FILE):
-            impact_exe = cfg.config.get_default(CONFIG_IMPACT_EXE_FILE)
-            if not os.path.isabs(impact_exe) and impact_exe.startswith(".") and (cfg.config_path != None):
-                impact_exe = os.path.abspath(os.path.join(os.path.dirname(cfg.config_path), impact_exe))
+        if self.config.has_default(CONFIG_IMPACT_EXE_FILE):
+            impact_exe = self.config.get_default(CONFIG_IMPACT_EXE_FILE)
             return impact_exe
 
         return _DEFAULT_IMPACT_EXE
 
-    def build(self):
-        """Prepare working directory, execute IMPACT and read the resulting model.
+    def run(self):
+        """Prepare working directory and execute IMPACT.
         """
 
         data_dir = self.data_dir
-        if (data_dir == None) and cfg.config.has_default(CONFIG_IMPACT_DATA_DIR):
-            data_dir = cfg.config.get_default(CONFIG_IMPACT_DATA_DIR)
+        if (data_dir == None) and self.config.has_default(CONFIG_IMPACT_DATA_DIR):
+            data_dir = self.config.get_default(CONFIG_IMPACT_DATA_DIR)
 
         if data_dir == None:
-            raise RuntimeError("ResultFactory: No data directory provided, check the configuration")
+            raise RuntimeError("LatticeRunner: No data directory provided, check the configuration")
 
         work_dir = self.work_dir
 
         impact_exe = self._get_config_impact_exe()
 
         if not os.path.isdir(data_dir):
-            raise RuntimeError("ResultFactory: Data directory not found: {}".format(data_dir))
+            raise RuntimeError("LatticeRunner: Data directory not found: {}".format(data_dir))
 
-        if (work_dir != None) and os.path.exists(work_dir):
-            raise RuntimeError("ResultFactory: Working directory already exists: {}".format(work_dir))
+        if (work_dir != None) and os.path.isfile(work_dir):
+            raise RuntimeError("LatticeRunner: Working directory must be a directory: {}".format(work_dir))
 
         if work_dir != None:
-            os.makedirs(work_dir)
+            if not os.path.exists(work_dir):
+                os.makedirs(work_dir)
             rm_work_dir = False
         else:
             work_dir = tempfile.mkdtemp(_TEMP_DIRECTORY_SUFFIX)
             rm_work_dir = True
 
-        _LOGGER.info("ResultFactory: Working directory: %s", work_dir)
+        _LOGGER.info("LatticeRunner: Working directory: %s", work_dir)
 
         if os.path.isabs(data_dir):
             abs_data_dir = data_dir
@@ -155,8 +167,9 @@ class ResultFactory(object):
             srcpath = os.path.join(abs_data_dir, datafile)
             destpath = os.path.join(work_dir, datafile)
             if os.path.isfile(os.path.join(abs_data_dir, datafile)):
-                os.symlink(srcpath, destpath)
-                _LOGGER.debug("ResultFactory: Link data file %s to %s", srcpath, destpath)
+                if not os.path.exists(destpath):
+                    os.symlink(srcpath, destpath)
+                    _LOGGER.debug("LatticeRunner: Link data file %s to %s", srcpath, destpath)
 
         try:
             with open(os.path.join(work_dir, "test.in"), "w") as fp:
@@ -174,11 +187,7 @@ class ResultFactory(object):
             self._remove_work_dir(work_dir, rm_work_dir)
             raise
 
-        try:
-            return build_result(impact="FRIB", directory=work_dir, keep=True)
-        finally:
-            self._remove_work_dir(work_dir, rm_work_dir)
-
+        return work_dir
 
 
     @staticmethod
@@ -186,7 +195,7 @@ class ResultFactory(object):
         """Cleanup the working directory.
         """
         if rm_work_dir:
-            _LOGGER.debug("ResultFactory: Cleanup: remove work directory")
+            _LOGGER.debug("LatticeRunner: Cleanup: remove work directory")
             shutil.rmtree(work_dir)
